fix(BOS-82): reconnect persistent subscription on acknowledgement failure
When eventstore connection is closed (due tue timeout or some other reason),
there is a chance that PersistentSubscription.acknowledge will throw a NullPointerException
because of a missing connection to eventstore.
To resolve this, we wrap all exceptions which are thrown from acknowledge method in a custom
EventStoreException. This ensures that persisten subscription will reconnect in onClose handler.